+++ src/views/Composer.js
@@ -1,10 +1,41 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 import { View, Text, TextInput, StyleSheet, TouchableHighlight } from 'react-native';
 import { useObservableState } from 'observable-hooks';
 import { colors } from '../constants';
+import Icon from './components/Icon';
+import externalService from '../services/external';
 
-export default function Composer({ room }) {
+export default function Composer({ room, isEditing, onEndEdit, selectedMessage }) {
   const [value, setValue] = useState('');
+
+  const textInputRef = useRef(null);
+
+  const roomName = useObservableState(room.name$);
+
+  const handleSend = () => {
+    room.sendMessage(value, 'm.text');
+    setValue('');
+  };
+
+  const cancelEdit = () => {
+    setValue('');
+    onEndEdit();
+  };
+
+  const confirmEdit = () => {
+    externalService.editMessage(room.id, selectedMessage.id, value);
+    setValue('');
+    textInputRef.current.blur();
+    onEndEdit();
+  };
+
+  useEffect(() => {
+    if (isEditing && selectedMessage) {
+      console.log(selectedMessage.content$.getValue());
+      setValue(selectedMessage.content$.getValue().text);
+      textInputRef.current.focus();
+    }
+  }, [isEditing, selectedMessage]);
 
   if (!room) {
     return (
@@ -14,29 +45,51 @@ export default function Composer({ room }) {
     );
   }
 
-  const roomName = useObservableState(room.name$);
-
-  const handleSend = () => {
-    room.sendMessage(value, 'm.text');
-    setValue('');
-  };
-
   return (
     <View style={styles.wrapper}>
-      <TextInput
-        style={styles.input}
-        multiline
-        placeholder={`Message ${roomName}...`}
-        value={value}
-        onChangeText={setValue}
-      />
-      <TouchableHighlight
-        disabled={value.length === 0}
-        onPress={handleSend}
-        underlayColor="#ddd"
-        style={styles.sendButton}>
-        <Text style={[styles.sendText, value.length === 0 ? { color: '#888' } : {}]}>Send</Text>
-      </TouchableHighlight>
+      {isEditing && (
+        <View
+          style={{
+            margin: 6,
+            padding: 6,
+            borderLeftWidth: 4,
+            borderLeftColor: 'dodgerblue',
+            flexDirection: 'row',
+            alignItems: 'center',
+            justifyContent: 'space-between',
+          }}>
+          <View>
+            <Text style={{ color: 'dodgerblue', fontWeight: 'bold' }}>Editing</Text>
+            <Text numberOfLines={1} style={{ color: 'gray' }}>
+              {selectedMessage.content$?.getValue()?.text}
+            </Text>
+          </View>
+          <TouchableHighlight onPress={cancelEdit} underlayColor="#ddd">
+            <View>
+              <Icon name="close" color="gray" />
+            </View>
+          </TouchableHighlight>
+        </View>
+      )}
+      <View style={{ flexDirection: 'row', alignItems: 'flex-end' }}>
+        <TextInput
+          ref={textInputRef}
+          style={styles.input}
+          multiline
+          placeholder={`Message ${roomName}...`}
+          value={value}
+          onChangeText={setValue}
+        />
+        <TouchableHighlight
+          disabled={value.length === 0}
+          onPress={isEditing ? confirmEdit : handleSend}
+          underlayColor="#ddd"
+          style={styles.sendButton}>
+          <Text style={[styles.sendText, value.length === 0 ? { color: '#888' } : {}]}>
+            {isEditing ? 'Save' : 'Send'}
+          </Text>
+        </TouchableHighlight>
+      </View>
     </View>
   );
 }
@@ -46,8 +99,6 @@ const styles = StyleSheet.create({
     minHeight: 45,
     borderTopWidth: 1,
     borderTopColor: colors.gray300,
-    flexDirection: 'row',
-    alignItems: 'flex-end',
     backgroundColor: colors.white,
     padding: 6,
   },