+++ CHANGELOG.md
@@ -15,7 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- accidental console.log
+
 ### Fixed
+
+- keys needed for invite list
 
 ### Security